+++ src/pushsource/_impl/helpers.py
@@ -57,3 +57,31 @@ def try_int(value):
         return int(value)
     except ValueError:
         return value
+
+
+def try_bool(value):
+    """Convert values into booleans where possible.
+
+    This function is intended for use in :class:`~pushsource.Source` constructors
+    to ease the processing of arguments passed via URL (which always arrive as a string).
+
+    Parameters:
+        value
+            Any value.
+
+    Returns:
+        bool
+            ``value`` converted to a boolean, if it was a string representation
+            of a boolean.
+        other
+            ``value`` unmodified, in any other case.
+    """
+    if not isinstance(value, six.string_types):
+        return value
+
+    value = value.lower()
+    if value in ["1", "true", "yes"]:
+        return True
+    elif value in ["", "0", "false", "no"]:
+        return False
+    raise ValueError("Expected a boolean, got %s" % repr(value))